requirements.txt
aiogram>=3.0.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
